Derive the card from the targetted device
Unless explicitly specified, target a card based on the selected device.
Note that not all devices will be tied to a card.

falconindy  » put differently, if i have multiple cards, what determines which
              card is used by pulse for a given app?
tanuk       » In theory, the logic can be anything (it depends on what
              policy-implementing modules are loaded). By default, routing is mostly
              handled by module-stream-restore, which chooses the sink based on the
              user's previous routing choices.
tanuk       » If the user hasn't done any routing choices, the fallback logic is to
              select the current "default sink".
tanuk       » I don't recommend trying to guess the routing policy.
falconindy  » i guess my understanding of pulse internals is lacking
falconindy  » but that's rather enlightening
falconindy  » is there any way to figure out the connection between a sink and a card?
tanuk       » Yes... (One moment, I'll look up things.)
falconindy  » ah. uint32_t card
falconindy  » appears to be in pa_sink_info
falconindy  » so that ties the sink to the index of a card?
tanuk       » Yep.
falconindy  » awesome, that's good enough for what i need to do
tanuk       » Not all sinks are part of a card, though, but those that are will have
              the card index set.
falconindy  » also good to know