Don't connect with PA_CONTEXT_NOFAIL
NOFAIL means we don't fail if the daemon is not available when
pa_context_connect() is called and instead enter PA_CONTEXT_CONNECTING
state and wait for the daemon to appear.

So currently if pulseaudio isn't running or available, ponymix will just
block and wait. Since a lot of people use ponymix bound to their volume
keys, this could potentially mean spawning 100s of processes before
actually catching on that things aren't working.

PA_CONTEXT_NOFLAGS was introduced in 0.9.19. Its really just 0.

Signed-off-by: Simon Gomizelj <simongmzlj@gmail.com>